+++ test/util.test.ts
@@ -7,7 +7,6 @@ import {
   readFilters,
   readOutput,
   resultIsForTestCase,
-  transformOutput,
   varsMatch,
   writeMultipleOutputs,
   writeOutput,
@@ -366,51 +365,6 @@ describe('util', () => {
     const filters = await readFilters({ testFilter: 'filter.js' });
 
     expect(filters.testFilter).toBe(mockFilter);
-  });
-
-  describe('transformOutput', () => {
-    afterEach(() => {
-      jest.clearAllMocks();
-      jest.resetModules();
-    });
-
-    it('transforms output using a direct function', async () => {
-      const output = 'original output';
-      const context = { vars: { key: 'value' }, prompt: { id: '123' } };
-      const transformFunction = 'output.toUpperCase()';
-      const transformedOutput = await transformOutput(transformFunction, output, context);
-      expect(transformedOutput).toBe('ORIGINAL OUTPUT');
-    });
-
-    it('transforms output using an imported function from a file', async () => {
-      const output = 'hello';
-      const context = { vars: { key: 'value' }, prompt: { id: '123' } };
-      jest.doMock(path.resolve('transform.js'), () => (output: string) => output.toUpperCase(), {
-        virtual: true,
-      });
-      const transformFunctionPath = 'file://transform.js';
-      const transformedOutput = await transformOutput(transformFunctionPath, output, context);
-      expect(transformedOutput).toBe('HELLO');
-    });
-
-    it('throws error if transform function does not return a value', async () => {
-      const output = 'test';
-      const context = { vars: {}, prompt: {} };
-      const transformFunction = ''; // Empty function, returns undefined
-      await expect(transformOutput(transformFunction, output, context)).rejects.toThrow(
-        'Transform function did not return a value',
-      );
-    });
-
-    it('throws error if file does not export a function', async () => {
-      const output = 'test';
-      const context = { vars: {}, prompt: {} };
-      jest.doMock(path.resolve('transform.js'), () => 'banana', { virtual: true });
-      const transformFunctionPath = 'file://transform.js';
-      await expect(transformOutput(transformFunctionPath, output, context)).rejects.toThrow(
-        'Transform transform.js must export a function or have a default export as a function',
-      );
-    });
   });
 
   describe('providerToIdentifier', () => {
